Added Mediator for Username/Password anonymity
The motif to add this file was to encourage tinkerers to keep their login credentials anonymous while testing/contributing. We highly discourage the use of username and password fields in plainText as it can have severe security ramifications.
A future analogue/port/fix for Mac and Linux is highly recommended.
Kindly merge, Admin.